myproject/requirements.txt
gunicorn==21.2.0
packaging==23.1
psycopg2==2.9.7

asgiref==3.7.2
Django==5.0.2
django-cors-headers==4.3.1
djangorestframework==3.14.0
pytz==2024.1
sqlparse==0.4.4
typing_extensions==4.9.0
tzdata==2023.4

whitenoise==6.5.0
rapidminer==10.1.0